% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_covariates.R
\name{build_covariates}
\alias{build_covariates}
\title{Simulate a set of covariates}
\usage{
build_covariates(d, n, cdf = "unif")
}
\arguments{
\item{d}{Covariate dimension}

\item{n}{Sample size}

\item{cdf}{Covariate distribution, default to U(0,1)}
}
\value{
A n by d matrix
}
\description{
Simulate a set of covariates
}
\examples{
X <- build_covariates(10, 1000)

}
\author{
Jiacheng He
}
