% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphical_illustration.R
\name{graphical_illustration}
\alias{graphical_illustration}
\title{Visualize the replication}
\usage{
graphical_illustration(DGP, d, n_train, n_test)
}
\description{
This function is used to generate the plot in the "An Example" Section of the slide
}
\author{
Jiacheng He
}
