% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate_sim.R
\name{replicate_sim}
\alias{replicate_sim}
\title{Run one simulation of causal forest}
\usage{
replicate_sim(d = 10, n_train = 1000, n_test = 1000, DGP = DGP_1,
  sample.fraction = 0.5, mtry = ceiling(2 * d/3), num.trees = 2000,
  min.node.size = NULL, lambda = 0, cf_seed = 904)
}
\arguments{
\item{d}{Covariate dimension}

\item{n_train}{Sample size of the training set}

\item{n_test}{Sample size of the test set}

\item{DGP}{Which data generating process to adopt}

\item{sample.fraction}{A parameter passed to causal_forest()}

\item{mtry}{A parameter passed to causal_forest()}

\item{num.trees}{A parameter passed to causal_forest()}

\item{min.node.size}{A parameter passed to causal_forest()}

\item{lambda}{A parameter passed to causal_forest()}

\item{cf_seed}{A parameter passed to causal_forest()}
}
\value{
The performance evaluation of the causal forest of this simulation (MSE and coverage rate)
}
\description{
This function accepts the key parameters and set up one simulation
}
\examples{

}
\author{
Jiacheng He
}
