% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulation}
\alias{simulation}
\title{The function accepts the key parameters to execute everything}
\usage{
simulation(param_list, param_type, DGP, bs_num = 100, n_test = 100,
  evaluate = FALSE, file_name = NULL)
}
\arguments{
\item{param_list}{The values of the parameter we want to vary}

\item{param_type}{The name of the parameter in the causal_forest()}

\item{DGP}{The list of data generating process}

\item{bs_num}{How many times of replication for each senario}

\item{n_test}{Sample size of the test set}

\item{file_name}{The name of the file we want to the save the results in}
}
\value{
A data frame containing the MSE and coverage rate for each simulation under different parameter
settings
}
\description{
On the final script (which might be submitted to the cluster) we only run this function
}
\examples{
result_n_DGP1 <- simulation(param_list = n_list, param_type = "n", file_name = "output_n_DGP1")
}
\author{
Jiacheng He
}
