% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cf.R
\name{test_cf}
\alias{test_cf}
\title{Evaluate the causal forest fitting}
\usage{
test_cf(cf, df_test)
}
\arguments{
\item{cf}{A causal forest object}

\item{df_test}{A test set data frame which we evaluate the causal forest on}

\item{alpha}{The p value of the confidence interval}
}
\value{
A list of MSE and coverage rate
}
\description{
This function accepts a causal forest fitting object and calculate the MSE and coverage rate
}
\examples{
evaluate_cf(cf_1, test_set)
}
\author{
Jiacheng He
}
