% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_cf.R
\name{train_cf}
\alias{train_cf}
\title{Train causal forest}
\usage{
train_cf(DGP, param, param_type, cf_seed = 904)
}
\arguments{
\item{DGP}{Which data generating process to adopt}

\item{cf_seed}{A parameter passed to causal_forest()}

\item{d}{Covariate dimension}

\item{n_train}{Sample size of the training set}

\item{sample.fraction}{A parameter passed to causal_forest()}

\item{mtry}{A parameter passed to causal_forest()}

\item{num.trees}{A parameter passed to causal_forest()}

\item{min.node.size}{A parameter passed to causal_forest()}

\item{lambda}{A parameter passed to causal_forest()}
}
\value{
The performance evaluation of the causal forest of this simulation (MSE and coverage rate)
}
\description{
This function accepts the key parameters and set up one training simulation
}
\examples{

}
\author{
Jiacheng He
}
