% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_cf.R
\name{evaluate_cf}
\alias{evaluate_cf}
\title{Evaluate the MSE and coverage rate}
\usage{
evaluate_cf(df_fit, alpha = 0.05)
}
\description{
This function receives the point predictions and true values
and compute the MSE and confidence coverage rate
}
\author{
Jiacheng He
}
