% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_true_model.R
\name{generate_true_model}
\alias{generate_true_model}
\title{Generate the components of the variables}
\usage{
generate_true_model(X, DGP_tau, DGP_e, DGP_m, noise_sd = 1)
}
\arguments{
\item{X}{The covariate matrix}

\item{DGP_tau}{An expression, the DGP for tau}

\item{DGP_e}{An expression, the DGP for e}

\item{DGP_m}{An expression, the DGP for m}

\item{noise_sd}{The standard deviation of the Gaussian noise of the model}
}
\value{
A list of X, Y, W, tau, e, and m. Prepare for the subsequent use of model fitting and evaluation.
}
\description{
This function accepts the baseline covariates and data generating process,
then return tau, m, and e, then simulate Y
}
\examples{
data <- generate_true_model(X, DGP1$tau, DGP1$e, DGP$m)

}
\author{
Jiacheng He
}
